feat: extract ory API key into separate server-only file
