+++ .env
@@ -2,7 +2,8 @@ NEXT_PUBLIC_ORY_CONSOLE_URL=https://console.ory.sh
 NEXT_PUBLIC_ORY_PROJECTS_URL=https://project.console.ory.sh
 NEXT_PUBLIC_ORY_WEB_URL=https://ory.sh
 NEXT_PUBLIC_ORY_API_URL=https://api.console.ory.sh
-HUBSPOT_API_KEY=
-ORY_API_KEY=
 NEXT_PUBLIC_MATOMO_URL=https://sqa-web.ory.sh
 NEXT_PUBLIC_MATOMO_SITE_ID=4
+
+CONFIDENTIAL_HUBSPOT_API_KEY=
+CONFIDENTIAL_ORY_API_KEY=
