+++ src/app/ory/getLogoutUrl.ts
@@ -1,4 +1,4 @@
-import { frontendApi as oryFrontendApi } from "./api"
+import { frontendApi as oryFrontendApi } from "./frontendApi"
 
 export const getLogoutUrl = async (returnTo?: string) => {
   const { data: session } = await oryFrontendApi.createBrowserLogoutFlow({
